[PAGOPA-627] get call: updated dockerfile and swagger
