+++ src/main/java/it/gov/pagopa/afm/utils/task/MarketPlaceTrigger.java
@@ -1,9 +1,7 @@
 package it.gov.pagopa.afm.utils.task;
 
 import it.gov.pagopa.afm.utils.service.MarketPlaceClient;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 public class MarketPlaceTrigger implements Runnable {
 
     private final MarketPlaceClient marketPlaceClient;
