+++ src/test/java/it/gov/pagopa/afm/utils/ImportCDIFunctionTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.lenient;
 
 import java.util.Arrays;
 import java.util.List;
@@ -14,6 +15,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import feign.FeignException;
@@ -21,6 +23,7 @@ import it.gov.pagopa.afm.utils.entity.CDI;
 import it.gov.pagopa.afm.utils.entity.StatusType;
 import it.gov.pagopa.afm.utils.model.bundle.BundleResponse;
 import it.gov.pagopa.afm.utils.model.bundle.CDIWrapper;
+import it.gov.pagopa.afm.utils.repository.CDICollectionRepository;
 import it.gov.pagopa.afm.utils.service.CDIService;
 import it.gov.pagopa.afm.utils.service.MarketPlaceClient;
 import reactor.core.publisher.Mono;
@@ -32,22 +35,28 @@ class ImportCDIFunctionTest {
 	private MarketPlaceClient marketPlaceClient;
 	
 	@Mock
+	private CDICollectionRepository cdisRepository;
+	
+	@Spy
 	private CDIService cdiService;
 	
 	@Test
 	void applyTest(){
 		// precondition
+		cdiService.setCdisRepository(cdisRepository);
+		cdiService.setMarketPlaceClient(marketPlaceClient);
 		BundleResponse response = BundleResponse.builder().idBundle("12345").build();
 		Mockito.when(marketPlaceClient.createBundleByList(eq("201"), any())).thenReturn(Arrays.asList(response));
-		
 		EasyRandom generator = new EasyRandom();
 		CDI cdi = generator.nextObject(CDI.class);
 		cdi.setIdPsp("201");
 		cdi.setValidityDateFrom("2022-12-15");
 		cdi.setCdiStatus(StatusType.NEW);
+		Mockito.when(cdisRepository.getWorkableCDIs()).thenReturn(Arrays.asList(new CDI[] {cdi}));
+		
 		
 		CDIWrapper input = CDIWrapper.builder().cdiItems(Arrays.asList(new CDI[] {cdi})).build();
-		Mono<List<BundleResponse>> responses = new ImportCDIFunction(marketPlaceClient, cdiService).apply(Mono.just(input));
+		Mono<List<BundleResponse>> responses = new ImportCDIFunction(cdiService).apply(Mono.just(input));
 		
 		assertTrue(responses.block().size() > 0);
 		assertThat(responses.block().get(0).getIdBundle()).isEqualTo("12345");
@@ -56,16 +65,18 @@ class ImportCDIFunctionTest {
 	@Test
 	void applyTest_400(){
 		// precondition
+		cdiService.setCdisRepository(cdisRepository);
+		cdiService.setMarketPlaceClient(marketPlaceClient);
 		EasyRandom generator = new EasyRandom();
 		CDI cdi = generator.nextObject(CDI.class);
 		cdi.setIdPsp("400");
 		cdi.setValidityDateFrom("2022-12-15");
 		cdi.setCdiStatus(StatusType.NEW);
-		Mockito.when(marketPlaceClient.createBundleByList(eq("400"), any())).thenThrow(FeignException.BadRequest.class);
-		Mockito.when(cdiService.updateCDI(any(CDI.class))).thenReturn(cdi);
+		Mockito.when(cdisRepository.getWorkableCDIs()).thenReturn(Arrays.asList(new CDI[] {cdi}));
+		lenient().when(marketPlaceClient.createBundleByList(eq("400"), any())).thenThrow(FeignException.BadRequest.class);
 
 		CDIWrapper input = CDIWrapper.builder().cdiItems(Arrays.asList(new CDI[] {cdi})).build();
-		Mono<List<BundleResponse>> responses = new ImportCDIFunction(marketPlaceClient, cdiService).apply(Mono.just(input));
+		Mono<List<BundleResponse>> responses = new ImportCDIFunction(cdiService).apply(Mono.just(input));
 		
 		assertEquals(0,responses.block().size());
 	}
@@ -73,17 +84,19 @@ class ImportCDIFunctionTest {
 	@Test
 	void applyTest_409(){
 		// precondition
+		cdiService.setCdisRepository(cdisRepository);
+		cdiService.setMarketPlaceClient(marketPlaceClient);
 		EasyRandom generator = new EasyRandom();
 		CDI cdi = generator.nextObject(CDI.class);
 		cdi.setIdPsp("409");
 		cdi.setValidityDateFrom("2022-12-15");
 		cdi.setCdiStatus(StatusType.NEW);
-		Mockito.when(marketPlaceClient.createBundleByList(eq("409"), any())).thenThrow(FeignException.Conflict.class);
-		Mockito.when(cdiService.updateCDI(any(CDI.class))).thenReturn(cdi);
+		Mockito.when(cdisRepository.getWorkableCDIs()).thenReturn(Arrays.asList(new CDI[] {cdi}));
+		lenient().when(marketPlaceClient.createBundleByList(eq("409"), any())).thenThrow(FeignException.Conflict.class);
 		
 		
 		CDIWrapper input = CDIWrapper.builder().cdiItems(Arrays.asList(new CDI[] {cdi})).build();
-		Mono<List<BundleResponse>> responses = new ImportCDIFunction(marketPlaceClient, cdiService).apply(Mono.just(input));
+		Mono<List<BundleResponse>> responses = new ImportCDIFunction(cdiService).apply(Mono.just(input));
 		
 		assertEquals(0,responses.block().size());
 	}
@@ -91,17 +104,19 @@ class ImportCDIFunctionTest {
 	@Test
 	void applyTest_404(){
 		// precondition
+		cdiService.setCdisRepository(cdisRepository);
+		cdiService.setMarketPlaceClient(marketPlaceClient);
 		EasyRandom generator = new EasyRandom();
 		CDI cdi = generator.nextObject(CDI.class);
 		cdi.setIdPsp("404");
 		cdi.setValidityDateFrom("2022-12-15");
 		cdi.setCdiStatus(StatusType.NEW);
-		Mockito.when(marketPlaceClient.createBundleByList(eq("404"), any())).thenThrow(FeignException.NotFound.class);
-		Mockito.when(cdiService.updateCDI(any(CDI.class))).thenReturn(cdi);
+		Mockito.when(cdisRepository.getWorkableCDIs()).thenReturn(Arrays.asList(new CDI[] {cdi}));
+		lenient().when(marketPlaceClient.createBundleByList(eq("404"), any())).thenThrow(FeignException.NotFound.class);
 		
 		
 		CDIWrapper input = CDIWrapper.builder().cdiItems(Arrays.asList(new CDI[] {cdi})).build();
-		Mono<List<BundleResponse>> responses = new ImportCDIFunction(marketPlaceClient, cdiService).apply(Mono.just(input));
+		Mono<List<BundleResponse>> responses = new ImportCDIFunction(cdiService).apply(Mono.just(input));
 		
 		assertEquals(0,responses.block().size());
 	}
