+++ src/test/java/it/gov/pagopa/afm/utils/controller/ImportCDIControllerTest.java
@@ -1,0 +1,48 @@
+package it.gov.pagopa.afm.utils.controller;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import java.util.ArrayList;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+
+import it.gov.pagopa.afm.utils.service.CDIService;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+class ImportCDIControllerTest {
+		
+	@Autowired
+    private MockMvc mvc;
+	
+	@Mock
+	private CDIService cdiService;
+	
+	@BeforeEach
+    void setUp() {
+		// precondition
+        when(cdiService.syncCDI()).thenReturn(new ArrayList<>());
+    }
+
+	@Test
+    void syncCDI() throws Exception {
+        String url = "/cdis/sync";
+        MvcResult result = mvc.perform(get(url).contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+                .andReturn();
+        assertNotNull(result.getResponse().getContentAsString());
+    }
+}
