+++ src/test/java/it/gov/pagopa/afm/utils/service/CDIServiceTest.java
@@ -14,18 +14,27 @@ import it.gov.pagopa.afm.utils.repository.CDICollectionRepository;
 class CDIServiceTest {
 	@Mock
     private CDICollectionRepository cdisRepository;
+	@Mock
+	private MarketPlaceClient marketPlaceClient;
 	
 	@Test
 	void deleteTest() {
-		CDIService cdiService = new CDIService(cdisRepository);
+		CDIService cdiService = new CDIService(cdisRepository, marketPlaceClient);
 		cdiService.deleteCDI(CDI.builder().build());
 		assertTrue(true);
 	}
 	
 	@Test
 	void updateTest() {
-		CDIService cdiService = new CDIService(cdisRepository);
+		CDIService cdiService = new CDIService(cdisRepository, marketPlaceClient);
 		cdiService.updateCDI(CDI.builder().build());
 		assertTrue(true);
 	}
+	
+	@Test
+	void getWorkableCDIs() {
+		CDIService cdiService = new CDIService(cdisRepository, marketPlaceClient);
+		cdiService.getWorkableCDIs();
+		assertTrue(true);
+    }
 }
